add chart release workflow
Added a GitHub workflow that activates when something is pushed to main
or any baranch with "releases/" prefix.
The workflow uses the chart releaser GitHub action. It scans for new versions,
if any creates GitHub releases, and finally updates the index.yaml in gh-pages
branch, efectively turning the repository into a helm repository.